new rule called build_shapes
